[fix] 変数の命名やロジックの移植
